Reworking convertion logic
Pulling convert_pgmpy_to_conin logic into the
conin.common.pgmpy module.
