+++ conin/bayesian_network/__init__.py
@@ -5,4 +5,3 @@ from .model import (
     DiscreteBayesianNetwork,
     ConstrainedDiscreteBayesianNetwork,
 )
-from .model_pgmpy import convert_to_DiscreteBayesianNetwork
