+++ conin/common/pgmpy/__init__.py
@@ -1,6 +1,7 @@
 from .load_model import load_model
 from .log_potential import log_potential
 from .polytree import is_polytree
+from .mapcpd import MapCPD
 from .to_pgmax import convert_pgmpy_to_pgmax
 from .to_pomegranate import convert_pgmpy_to_pomegranate
-from .mapcpd import MapCPD
+from .to_conin import convert_pgmpy_to_conin
